rag2025/requirements.txt
# Core ML and NLP libraries (verified working versions)
numpy==2.1.0
sentence-transformers>=3.0.0
rank-bm25==0.2.2

# Web framework  
fastapi>=0.100.0
uvicorn[standard]>=0.24.0

# Data validation and serialization (upgraded to v2 for Python 3.13 compatibility)
pydantic>=2.0.0
pydantic-settings>=2.0.0
jsonschema==4.17.3

# Configuration and data handling
pyyaml==6.0.1
chardet==5.2.0

# Tokenization (try installing without tiktoken first)
# tiktoken==0.4.0

# Logging
loguru==0.7.2

# Testing (optional - comment out if not needed)
# pytest==7.4.3 
# pytest-asyncio==0.21.1

# Development (optional)
black==23.11.0

# Optional dependencies for testing and scripts
httpx==0.25.2
requests==2.31.0

# ========== RAG 2025 UPDATES ==========

# Vector Database
qdrant-client>=1.7.0

# LLM Providers
google-generativeai>=0.3.0
groq>=0.4.0
openai>=1.0.0  # For GLM-4 via Z.AI endpoint
zai-sdk>=0.1.0  # GLM-4.5 via Z.AI

# BGE Multi-Vector (AITeamVN)
transformers>=4.35.0
torch>=2.0.0

# Additional utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0
tenacity>=8.2.0  # For retry logic
tqdm>=4.66.0  # Progress bars
